Move EditVisibilityForm into gallery/forms.py
